Removed unprintable characters

+++ org.zikula.modulestudio.generator/src/org/zikula/modulestudio/generator/cartridges/zclassic/view/pages/Delete.xtend
@@ -25,79 +25,79 @@ class Delete {
     }
 
     def private deleteView(Entity it, String appName) '''
-        «val app = container.application»
-        {* purpose of this template: «nameMultiple.formatForDisplay» delete confirmation view *}
+        val app = container.application
+        {* purpose of this template: nameMultiple.formatForDisplay delete confirmation view *}
         {assign var='lct' value='user'}
         {if isset($smarty.get.lct) && $smarty.get.lct eq 'admin'}
             {assign var='lct' value='admin'}
         {/if}
-        «IF app.targets('1.3.5')»
+        IF app.targets('1.3.5')
             {include file="`$lct`/header.tpl"}
-        «ELSE»
+        ELSE
             {assign var='lctUc' value=$lct|ucfirst}
             {include file="`$lctUc`/header.tpl"}
-        «ENDIF»
-        <div class="«appName.toLowerCase»-«name.formatForDB» «appName.toLowerCase»-delete">
-            {gt text='Delete «name.formatForDisplay»' assign='templateTitle'}
+        ENDIF
+        <div class="appName.toLowerCase-name.formatForDB appName.toLowerCase-delete">
+            {gt text='Delete name.formatForDisplay' assign='templateTitle'}
             {pagesetvar name='title' value=$templateTitle}
-            «templateHeader»
+            templateHeader
 
-            <p class="«IF app.targets('1.3.5')»z-warningmsg«ELSE»alert alert-warningmsg«ENDIF»">{gt text='Do you really want to delete this «name.formatForDisplay» ?'}</p>
+            <p class="IF app.targets('1.3.5')z-warningmsgELSEalert alert-warningmsgENDIF">{gt text='Do you really want to delete this name.formatForDisplay ?'}</p>
 
-            <form class="«IF app.targets('1.3.5')»z-form«ELSE»form-horizontal«ENDIF»" action="«IF app.targets('1.3.5')»{modurl modname='«appName»' type=$lct func='delete' ot='«name.formatForCode»' «routeParamsLegacy(name, true, false)»}«ELSE»{route name='«appName.formatForDB»_«name.formatForCode»_delete' «routeParams(name, true)» lct=$lct}«ENDIF»" method="post"«IF !app.targets('1.3.5')» role="form"«ENDIF»>
+            <form class="IF app.targets('1.3.5')z-formELSEform-horizontalENDIF" action="IF app.targets('1.3.5'){modurl modname='appName' type=$lct func='delete' ot='name.formatForCode' routeParamsLegacy(name, true, false)}ELSE{route name='appName.formatForDB_name.formatForCode_delete' routeParams(name, true) lct=$lct}ENDIF" method="post"IF !app.targets('1.3.5') role="form"ENDIF>
                 <div>
                     <input type="hidden" name="csrftoken" value="{insert name='csrftoken'}" />
                     <input type="hidden" id="confirmation" name="confirmation" value="1" />
                     <fieldset>
                         <legend>{gt text='Confirmation prompt'}</legend>
-                        <div class="«IF app.targets('1.3.5')»z-buttons z-formbuttons«ELSE»form-group form-buttons«ENDIF»">
-                        «IF !app.targets('1.3.5')»
+                        <div class="IF app.targets('1.3.5')z-buttons z-formbuttonsELSEform-group form-buttonsENDIF">
+                        IF !app.targets('1.3.5')
                             <div class="col-lg-offset-3 col-lg-9">
-                        «ENDIF»
+                        ENDIF
                             {gt text='Delete' assign='deleteTitle'}
-                            {button src='14_layer_deletelayer.png' set='icons/small' text=$deleteTitle title=$deleteTitle class='«IF app.targets('1.3.5')»z-btred«ELSE»btn btn-danger«ENDIF»'}
-                            «IF app.targets('1.3.5')»
-                                <a href="{modurl modname='«appName»' type=$lct func='view' ot='«name.formatForCode»'}">{icon type='cancel' size='small' __alt='Cancel' __title='Cancel'} {gt text='Cancel'}</a>
-                            «ELSE»
-                                <a href="{route name='«appName.formatForDB»_«name.formatForCode»_view' lct=$lct}" class="btn btn-default" role="button"><span class="fa fa-times"></span> {gt text='Cancel'}</a>
-                            «ENDIF»
-                        «IF !app.targets('1.3.5')»
+                            {button src='14_layer_deletelayer.png' set='icons/small' text=$deleteTitle title=$deleteTitle class='IF app.targets('1.3.5')z-btredELSEbtn btn-dangerENDIF'}
+                            IF app.targets('1.3.5')
+                                <a href="{modurl modname='appName' type=$lct func='view' ot='name.formatForCode'}">{icon type='cancel' size='small' __alt='Cancel' __title='Cancel'} {gt text='Cancel'}</a>
+                            ELSE
+                                <a href="{route name='appName.formatForDB_name.formatForCode_view' lct=$lct}" class="btn btn-default" role="button"><span class="fa fa-times"></span> {gt text='Cancel'}</a>
+                            ENDIF
+                        IF !app.targets('1.3.5')
                             </div>
-                        «ENDIF»
+                        ENDIF
                         </div>
                     </fieldset>
 
-                    «callDisplayHooks(appName)»
+                    callDisplayHooks(appName)
                 </div>
             </form>
         </div>
-        «IF app.targets('1.3.5')»
+        IF app.targets('1.3.5')
             {include file="`$lct`/footer.tpl"}
-        «ELSE»
+        ELSE
             {include file="`$lctUc`/footer.tpl"}
-        «ENDIF»
+        ENDIF
     '''
 
     def private templateHeader(Entity it) '''
         {if $lct eq 'admin'}
-            «IF container.application.targets('1.3.5')»
+            IF container.application.targets('1.3.5')
                 <div class="z-admin-content-pagetitle">
                     {icon type='delete' size='small' __alt='Delete'}
                     <h3>{$templateTitle}</h3>
                 </div>
-            «ELSE»
+            ELSE
                 <h3>
                     <span class="fa fa-trash-o"></span>
                     {$templateTitle}
                 </h3>
-            «ENDIF»
+            ENDIF
         {else}
             <h2>{$templateTitle}</h2>
         {/if}
     '''
 
     def private callDisplayHooks(Entity it, String appName) '''
-        {notifydisplayhooks eventname='«appName.formatForDB».ui_hooks.«nameMultiple.formatForDB».form_delete' id="«FOR pkField : getPrimaryKeyFields SEPARATOR '_'»`$«name.formatForCode».«pkField.name.formatForCode»`«ENDFOR»" assign='hooks'}
+        {notifydisplayhooks eventname='appName.formatForDB.ui_hooks.nameMultiple.formatForDB.form_delete' id="FOR pkField : getPrimaryKeyFields SEPARATOR '_'`$name.formatForCode.pkField.name.formatForCode`ENDFOR" assign='hooks'}
         {foreach key='providerArea' item='hook' from=$hooks}
         <fieldset>
             <legend>{$hookName}</legend>
